Split main page to create chart component
