+++ ClimateExplorer.Visualiser/Shared/DataSetBrowser.razor
@@ -253,7 +253,7 @@
                         {
                             new DataSetLibraryEntry()
                             {
-                                Name = "CO₂ (Carbon Dioxide) adjusted parts per million",
+                                Name = "CO₂ (Carbon Dioxide) parts per million",
                                 SourceSeriesSpecifications = 
                                     new DataSetLibraryEntry.SourceSeriesSpecification[]
                                     {
@@ -266,7 +266,7 @@
                             },
                             new DataSetLibraryEntry()
                             {
-                                Name = "CH₄ (Methane) adjusted parts per billion",
+                                Name = "CH₄ (Methane) parts per billion",
                                 SourceSeriesSpecifications = 
                                     new DataSetLibraryEntry.SourceSeriesSpecification[]
                                     {
@@ -280,7 +280,7 @@
                             },
                             new DataSetLibraryEntry()
                             {
-                                Name = "N₂O (Nitrous Oxide) adjusted parts per billion",
+                                Name = "N₂O (Nitrous Oxide) parts per billion",
                                 SourceSeriesSpecifications = 
                                     new DataSetLibraryEntry.SourceSeriesSpecification[]
                                     {
