+++ ClimateExplorer.Visualiser/Shared/NavMenu.razor
@@ -12,6 +12,12 @@
         <BarMenu>
             <BarStart>
                 <BarItem>
+                    <BarLink To="">local charts</BarLink>
+                </BarItem>
+                <BarItem>
+                    <BarLink To="regionalandglobal">regional & global charts</BarLink>
+                </BarItem>
+                <BarItem>
                     <BarLink To="about">about</BarLink>
                 </BarItem>
                 <BarItem>
