+++ ClimateExplorer.Visualiser/Shared/SelectLocation.razor
@@ -16,7 +16,7 @@
                 </Field>
             }
             <Field>
-                <Button Clicked="@FireUseCurrentLocation" Color="Color.Primary">Use Your Location</Button>
+                <Button Clicked="@FireUseCurrentLocation" Color="Color.Primary">Near me</Button>
 
                 @if (BrowserLocationErrorMessage != null)
                 {
