+++ ClimateExplorer.Visualiser/UiModel/ChartViewModel.cs
@@ -1,0 +1,5 @@
+﻿namespace ClimateExplorer.Visualiser.UiModel;
+
+public class ChartViewModel
+{
+}
