+++ ClimateExplorer.Visualiser/UiModel/DataDownloadPackage.cs
@@ -1,0 +1,9 @@
+﻿using ClimateExplorer.Core.DataPreparation;
+
+namespace ClimateExplorer.Visualiser.UiModel;
+
+public class DataDownloadPackage
+{
+    public List<SeriesWithData> ChartSeriesWithData { get; set; }
+    public BinIdentifier[] Bins { get; set; }
+}
